pyproject.toml
[project]
name = "fun-zotero-research-gpt"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Takfes", email = "fpgmaas@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.9",
    "langchain>=0.3.9",
    "langchain-openai>=0.2.11",
    "langchain-pinecone>=0.2.0",
    "langchainhub>=0.1.21",
    "openai>=1.57.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pyzotero>=1.5.28",
    "streamlit>=1.40.2",
    "langchain-ollama>=0.2.1",
    "pypdf>=5.1.0",
    "cohere>=5.13.3",
    "chromadb>=0.5.23",
    "langchain-cohere>=0.3.3",
    "pyprojroot>=0.3.0",
    "langchain-chroma>=0.1.4",
    "streamlit-chat>=0.1.1",
    "tqdm>=4.67.1",
    "streamlit-aggrid>=1.0.5",
]

[project.urls]
Homepage = "https://Takfes.github.io/fun-zotero-research-gpt/"
Repository = "https://github.com/Takfes/fun-zotero-research-gpt"
Documentation = "https://Takfes.github.io/fun-zotero-research-gpt/"

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "mypy>=0.991",

    "ruff>=0.6.9",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["zotgpt"]
# packages = "find:"
package-dir = { "" = "src" }
# py-modules = ["src"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 80
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Ignore unused import warnings
    "F401",
    # Ignore long messages outside the exception class warning
    "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
